+++ Data/Retail/ExportedAchievements.lua
@@ -39,7 +39,7 @@ local f = addon.Objects.Faction; -- Saves some characters and file size as we us
 -- [[ Everything after these lines is automatically generated as an export from ]] --
 -- [[ an SQLite database and is not meant for manual edit. - AUTOGENTOKEN ]] --
 
--- [[ Exported at 2024-04-27 15-10-43 ]] --
+-- [[ Exported at 2024-05-05 17-26-39 ]] --
 tasks = {
     {O, 40223, "From", "Event", "1495", "Until", "Event", "1495"},
     {N, 40223, 208}, -- Timerunner